+++ com.openexchange.metrics.micrometer/src/com/openexchange/metrics/micrometer/internal/filter/DistributionMinimumMicrometerFilterPerformer.java
@@ -79,6 +79,6 @@ public class DistributionMinimumMicrometerFilterPerformer extends AbstractMicrom
     @Override
     DistributionStatisticConfig applyConfig(Entry<String, String> entry, String metricId, DistributionStatisticConfig config) {
         Long value = distributionValueSanityCheck(metricId, entry.getValue());
-        return value == null ? config : DistributionStatisticConfig.builder().minimumExpectedValue(value.longValue()).build().merge(config);
+        return value == null ? config : DistributionStatisticConfig.builder().minimumExpectedValue(value).build().merge(config);
     }
 }
